+++ R/data.R
@@ -18,3 +18,9 @@ NULL
 #' @docType data
 NULL
 
+
+#' Mapping data for `disability` function
+#'
+#' @name mapping_disability
+#' @docType data
+NULL