+++ R/dummy_data.R
@@ -24,53 +24,58 @@
 #'
 #' @examples
 #' ### Simulate 100 ICD-10 codes based on the 2011 version.
-#' \dontrun{ sample_icd(100, source="comorbidity") }
+#' \dontrun{
+#' sample_icd(100, source = "comorbidity")
+#' }
 #'
 #' ### Simulate 100 ICD-10 codes starting with "C2" or "E10" based on the 2011 version.
-#' \dontrun{ sample_icd(100, source="comorbidity", pattern="^C2|^E10") }
+#' \dontrun{
+#' sample_icd(100, source = "comorbidity", pattern = "^C2|^E10")
+#' }
 #'
 #' ### Simulate 50 ICD-10-CA codes based on codes found in the `lookup_icd10_ca_description` table
 #' \dontrun{
 #' drv <- dbDriver("PostgreSQL")
 #' dbcon <- DBI::dbConnect(drv,
-#'                         dbname = "db",
-#'                         host = "172.XX.XX.XXX",
-#'                         port = 1234,
-#'                         user = getPass("Enter user:"),
-#'                         password = getPass("password"))
-#' sample_icd(50, source="icd_lookup", dbcon=dbcon)
+#'   dbname = "db",
+#'   host = "domain_name.ca",
+#'   port = 1234,
+#'   user = getPass("Enter user:"),
+#'   password = getPass("password")
+#' )
+#' sample_icd(50, source = "icd_lookup", dbcon = dbcon)
 #' }
 #'
 sample_icd <- function(n = 1, source = "comorbidity", dbcon = NULL, pattern = NULL) {
   switch(source,
-         comorbidity = {
-           comorb <- comorbidity::icd10_2011 %>% as.data.table()
-           if (!is.null(pattern)) {
-             comorb <- comorb[grepl(toupper(pattern), Code.clean)]
-           }
-           if (nrow(comorb) > 0) {
-             sample(x = comorb$Code.clean, size = n, replace = TRUE)
-           } else {
-             stop("No matching diagnoses found for the specified pattern")
-           }
-         },
-         icd_lookup = {
-           if (!is.null(dbcon)) {
-             lookup <- RPostgreSQL::dbGetQuery(dbcon, "SELECT diagnosis_code  FROM lookup_icd10_ca_description where type != 'category'") %>% as.data.table()
+    comorbidity = {
+      comorb <- comorbidity::icd10_2011 %>% as.data.table()
+      if (!is.null(pattern)) {
+        comorb <- comorb[grepl(toupper(pattern), Code.clean)]
+      }
+      if (nrow(comorb) > 0) {
+        sample(x = comorb$Code.clean, size = n, replace = TRUE)
+      } else {
+        stop("No matching diagnoses found for the specified pattern")
+      }
+    },
+    icd_lookup = {
+      if (!is.null(dbcon)) {
+        lookup <- RPostgreSQL::dbGetQuery(dbcon, "SELECT diagnosis_code  FROM lookup_icd10_ca_description where type != 'category'") %>% as.data.table()
 
-             if (!is.null(pattern)) {
-               lookup <- lookup[grepl(toupper(pattern), diagnosis_code)]
-             }
+        if (!is.null(pattern)) {
+          lookup <- lookup[grepl(toupper(pattern), diagnosis_code)]
+        }
 
-             if (nrow(lookup) > 0) {
-               sample(x = lookup$diagnosis_code, size = n, replace = TRUE)
-             } else {
-               stop("No matching diagnoses found for the specified pattern")
-             }
-           } else {
-             stop("Invalid input for 'dbcon' argument. Database connection is required for sampling from `lookup_icd10_ca_to_ccsr` table\n")
-           }
-         }
+        if (nrow(lookup) > 0) {
+          sample(x = lookup$diagnosis_code, size = n, replace = TRUE)
+        } else {
+          stop("No matching diagnoses found for the specified pattern")
+        }
+      } else {
+        stop("Invalid input for 'dbcon' argument. Database connection is required for sampling from `lookup_icd10_ca_to_ccsr` table\n")
+      }
+    }
   )
 }
 
@@ -130,35 +135,40 @@ sample_icd <- function(n = 1, source = "comorbidity", dbcon = NULL, pattern = NU
 #' @examples
 #'
 #' ### Simulate a erdiagnosis table for 5 unique subjects with total 20 records:
-#' \dontrun{ set.seed(1)
-#'           erdiag <- dummy_diag(nid=5, nrow=20, ipdiagnosis=F) }
+#' \dontrun{
+#' set.seed(1)
+#' erdiag <- dummy_diag(nid = 5, nrow = 20, ipdiagnosis = F)
+#' }
 #'
 #' ### Simulate a ipdiagnosis table with diagnosis codes starting with "E11":
-#' \dontrun{ set.seed(1)
-#'           ipdiag <- dummy_diag(nid=5, nrow=20, ipdiagnosis=T, pattern ="^E11") }
+#' \dontrun{
+#' set.seed(1)
+#' ipdiag <- dummy_diag(nid = 5, nrow = 20, ipdiagnosis = T, pattern = "^E11")
+#' }
 #'
 #' ### Simulate a ipdiagnosis table with random diagnosis codes in diagnosis type 3 or 6 only:
-#' \dontrun{ set.seed(1)
-#'           ipdiag <- dummy_diag(nid=5, nrow=20, diagnosis_type=(c("3", "6"))) %>%
-#'              filter(diagnosis_type!="M") # remove default rows with diagnosis_type="M" from each ID
+#' \dontrun{
+#' set.seed(1)
+#' ipdiag <- dummy_diag(nid = 5, nrow = 20, diagnosis_type = (c("3", "6"))) %>%
+#'   filter(diagnosis_type != "M") # remove default rows with diagnosis_type="M" from each ID
 #' }
 #'
 #' ### Simulate a ipdiagnosis table with ICD-10-CA codes:
 #' \dontrun{
 #' drv <- dbDriver("PostgreSQL")
 #' dbcon <- DBI::dbConnect(drv,
-#'                         dbname = "db",
-#'                         host = "172.XX.XX.XXX",
-#'                         port = 1234,
-#'                         user = getPass("Enter user:"),
-#'                         password = getPass("password"))
+#'   dbname = "db",
+#'   host = "172.XX.XX.XXX",
+#'   port = 1234,
+#'   user = getPass("Enter user:"),
+#'   password = getPass("password")
+#' )
 #'
 #' set.seed(1)
-#' ipdiag <- dummy_diag(nid=5, nrow=20, ipdiagnosis=T, dbcon=dbcon, source="icd_lookup")}
+#' ipdiag <- dummy_diag(nid = 5, nrow = 20, ipdiagnosis = T, dbcon = dbcon, source = "icd_lookup")
+#' }
 #'
-
 dummy_diag <- function(nid = 5, nrow = 50, ipdiagnosis = TRUE, diagnosis_type = NULL, ...) {
-
   df1 <- data.table(genc_id = 1:nid, diagnosis_type = "M") # ensure each id has a type M diagnosis
 
   if (!is.null(diagnosis_type)) {
@@ -170,8 +180,8 @@ dummy_diag <- function(nid = 5, nrow = 50, ipdiagnosis = TRUE, diagnosis_type = 
     df2 <- data.table(
       genc_id = sample(1:nid, size = (nrow - nid), replace = TRUE),
       diagnosis_type = sample(c("1", "2", "3", "4", "5", "6", "9", "W", "X", "Y"),
-                              size = (nrow - nid), replace = TRUE,
-                              prob = c(0.43, 0.07, 0.40, 0.005, 0.0002, 0.002, 0.07, 0.02, 0.0006, 0.00003)
+        size = (nrow - nid), replace = TRUE,
+        prob = c(0.43, 0.07, 0.40, 0.005, 0.0002, 0.002, 0.07, 0.02, 0.0006, 0.00003)
       )
     )
   }
@@ -184,9 +194,8 @@ dummy_diag <- function(nid = 5, nrow = 50, ipdiagnosis = TRUE, diagnosis_type = 
       diagnosis_prefix = sample(c("", "N", "Q", "6"), size = nrow, replace = TRUE, prob = c(0.9, 0.05, 0.02, 0.01))
     )
 
-  if (ipdiagnosis==FALSE) {
-
-    if(!is.null(diagnosis_type)){
+  if (ipdiagnosis == FALSE) {
+    if (!is.null(diagnosis_type)) {
       er_diagnosis_type <- sample(diagnosis_type, size = nrow, replace = TRUE)
     } else {
       er_diagnosis_type <- sample(c("", "M", "9", "3", "O"), size = nrow, replace = TRUE, prob = c(0.53, 0.38, 0.06, 0.02, 0.01))
@@ -199,6 +208,339 @@ dummy_diag <- function(nid = 5, nrow = 50, ipdiagnosis = TRUE, diagnosis_type = 
   }
 
   return(dummy[order(dummy$genc_id)])
+}
+
+
+
+#' @title
+#' Simulate ipadmdad data
+#'
+#' @description
+#' This function creates a dummy dataset with a subset of variables that
+#' are contained in the GEMINI "ipadmdad" table (see details in
+#' [GEMINI Data Repository Dictionary](https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5)).
+#'
+#' The simulated encounter-level variables that are returned by this function
+#' are currently: Admission date-time, discharge date-time, age, gender,
+#' discharge disposition, transfer to an alternate level of care (ALC), and ALC
+#' days. The distribution of these simulated variables roughly mimics the real
+#' distribution of each variable observed in the GIM cohort from 2015-2022.
+#' Admission date-time is simulated in conjunction with discharge date-time to
+#' mimic realistic length of stay. All other variables are simulated
+#' independently of each other, i.e., there is no correlation between age,
+#' gender, discharge disposition etc. that may exist in the real data. One
+#' exception to this is `number_of_alc_days`, which is only > 0 for entries
+#' where `alc_service_transfer_flag == TRUE` and the length of ALC is capped at
+#' the total length of stay.
+#'
+#' The function simulates patient populations that differ across hospitals. That
+#' is, patient characteristics are simulated separately for each hospital, with
+#' a different, randomly drawn distribution mean (i.e., random intercepts).
+#' However, the degree of hospital-level variation simulated by this function
+#' is arbitrary and does not reflect true differences between hospitals in the
+#' real GEMINI dataset.
+#'
+#' @param n (`integer`)\cr Total number of encounters (`genc_ids`) to be
+#' simulated.
+#'
+#' @param n_hospitals (`integer`)\cr
+#' Number of hospitals to be simulated. Total number of `genc_ids` will be split
+#' up pseudo-randomly between hospitals to ensure roughly equal sample size at
+#' each hospital.
+#'
+#' @param fisc_years (`numeric`)\cr
+#' A numeric vector containing the time period, specified as fiscal years
+#' (starting in April each year). For example, `c(2015, 2019)` generates data
+#' from 2015-04-01 to 2020-03-31.
+#'
+#' @return (`data.frame`)\cr A data.frame object similar to the "ipadmdad" table
+#' containing the following fields:
+#' - `genc_id` (`integer`): GEMINI encounter ID
+#' - `hospital_num` (`integer`): Hospital ID
+#' - `admission_date_time` (`character`): Date-time of admission in YYYY-MM-DD HH:MM format
+#' - `discharge_date_time` (`character`): Date-time of discharge in YYYY-MM-DD HH:MM format
+#' - `age` (`integer`): Patient age
+#' - `gender` (`character`): Patient gender (F/M/O for Female/Male/Other)
+#' - `discharge_disposition` (`integer`): All valid categories according to DAD
+#' abstracting manual 2022-2023
+#'    - 4: Home with Support/Referral
+#'    - 5: Private Home
+#'    - 8: Cadaveric Donor (does not exist in GEMINI data)
+#'    - 9: Stillbirth (does not exist in GEMINI data)
+#'    - 10: Transfer to Inpatient Care
+#'    - 20: Transfer to ED and Ambulatory Care
+#'    - 30: Transfer to Residential Care
+#'    - 40: Transfer to Group/Supportive Living
+#'    - 90: Transfer to Correctional Facility
+#'    - 61: Absent Without Leave (AWOL)
+#'    - 62: Left Against Medical Advice (LAMA)
+#'    - 65: Did not Return from Pass/Leave
+#'    - 66: Died While on Pass/Leave
+#'    - 67: Suicide out of Facility (does not exist in GEMINI data)
+#'    - 72: Died in Facility
+#'    - 73: Medical Assistance in Dying (MAID)
+#'    - 74: Suicide in Facility
+#' - `alc_service_transfer_flag` (`character`): Variable indicating whether
+#' patient was transferred to an alternate level of care (ALC) during their
+#' hospital stay. Coding is messy and varies across sites. Possible values are:
+#'    - Missing: `NA`, `""`
+#'    - True: `"TRUE"/"true"/"T"`, `"y"/"Y"`, `"1"/"99"`, `"ALC"`
+#'    - False: `"FALSE"/"false"`, `"N"`, `"0"`, `"non-ALC"`
+#' Some entries with missing `alc_service_transfer_flag` can be inferred based
+#' on value of `number_of_alc_days` (see below)
+#' - `number_of_alc_days` (`integer`): Number of days spent in ALC (rounded to
+#' nearest integer). If `number_of_alc_days = 0`, no ALC occurred;
+#' if `number_of_alc_days > 0`, ALC occurred.
+#' Note that days spent in ALC should usually be < length of
+#' stay. However, due to the fact that ALC days are rounded up, it's possible
+#' for `number_of_alc_days` to be larger than `los_days_derived`.
+#'
+#' @importFrom sn rsn
+#' @importFrom MCMCpack rdirichlet
+#' @importFrom lubridate ymd_hm
+#' @export
+#'
+#' @examples
+#' # Simulate 10,000 encounters from 10 hospitals for fiscal years 2018-2020.
+#' ipadmdad <- dummy_ipadmdad(n = 10000, n_hospitals = 10, time_period = c(2018, 2020))
+#'
+dummy_ipadmdad <- function(n = 1000,
+                           n_hospitals = 10,
+                           time_period = c(2015, 2023)) {
+  ############### CHECKS: Make sure n is at least n_hospitals * length(time_period)
+  if (n < n_hospitals * length(time_period)) {
+    stop("Invalid user input.
+    Number of encounters `n` should at least be equal to `n_hospitals` * `length(time_period)`")
+  }
+
+  ############### PREPARE OUTPUT TABLE ###############
+  ## create all combinations of hospitals and fiscal years
+  hospital_num <- seq(1, n_hospitals, 1)
+  year <- seq(time_period[1], time_period[2], 1)
+
+  data <- expand.grid(hospital_num = hospital_num, year = year) %>% data.table()
+
+  # randomly draw number of encounters per hospital*year combo
+  # make sure they add up to desired total number of encounters
+  data[, n := rmultinom(1, n, rep.int(1 / nrow(data), nrow(data)))]
+  sum(data$n)
+
+  # blow up row number according to encounter per combo
+  data <- data[rep(seq(nrow(data)), data$n), ]
+
+  # turn year variable into actual date by randomly drawing date_time
+  add_random_datetime <- function(year) {
+    start_date <- paste0(year, "-04-01 00:00 UTC") # start each fisc year on Apr 1
+    end_date <- paste0(year + 1, "-03-31 23:59 UTC") # end of fisc year
+
+    random_datetime <- format(
+      as.POSIXct(runif(length(year), as.POSIXct(start_date), as.POSIXct(end_date)),
+        origin = "1970-01-01"
+      ),
+      format = "%Y-%m-%d %H:%M"
+    )
+
+    return(random_datetime)
+  }
+
+  data[, discharge_date_time := add_random_datetime(year)]
+
+  # add genc_id from 1-n
+  data <- data[order(discharge_date_time), ]
+  data[, genc_id := seq(1, nrow(data), 1)]
+
+
+  ############### DEFINE VARIABLE DISTRIBUTIONS ###############
+  ## AGE
+  # create left-skewed distribution, truncated from 18-110
+  age_distr <- function(n = 10000, xi = 95, omega = 30, alpha = -10) {
+    age <- rsn(n, xi, omega, alpha)
+
+    # truncate at [18, 110]
+    age <- as.integer(age[age >= 18 & age <= 110])
+  }
+
+
+  ############### ADD VARIABLES CLUSTERED BY HOSPITAL ###############
+  # Any encounter characteristics (e.g., age/gender/discharge disposition) are
+  # simulated as being clustered by hospital (i.e., each hospital will be
+  # simulated as random intercept, i.e., different location parameter)
+  add_vars <- function(hosp_data) {
+    n_enc <- nrow(hosp_data)
+
+    ## AGE
+    # create new age distribution for each hospital where location parameter xi
+    # varies to create a random intercept by site
+    age <- age_distr(xi = rnorm(1, 95, 5))
+    hosp_data[, age := sample(age, n_enc, replace = TRUE)]
+
+    ## GENDER (F/M/Other)
+    prob <- data.table(
+      "gender" = c("F", "M", "O"),
+      "p" = c(.501, .498, 0.001 + 1e-5)
+    ) # add small constant to Os to ensure it's not rounded to 0 below
+    # Introduce random hospital-level variability
+    prob[, p := t(rdirichlet(1, alpha = prob$p / 0.005))] # 0.005 = level of variability
+    hosp_data[, gender := sample(prob$gender, n_enc, replace = TRUE, prob$p / sum(prob$p))] # make sure probs add up to 1 (see addition of constant above)
+
+    ## DISCHARGE DISPOSITION
+    prob <- data.table(
+      "discharge_disposition" = c(4, 5, 8, 9, 10, 20, 30, 40, 61, 62, 65, 66, 67, 72, 73, 74, 90),
+      "p" = c(.275, .386, 0, 0, .143, 0.002, .045, .040, .001 + 1e-5, .028, 0.001 + 1e-5, 0.001 + 1e-5, 0, .079, .001 + 1e-5, 0.001 + 1e-5, .001)
+    ) # add small constant to Os to ensure it's not rounded to 0 below
+    prob[, p := t(rdirichlet(1, alpha = prob$p / 0.005))] # 0.005 = level of hospital-level variability
+    hosp_data[, discharge_disposition := as.integer(sample(prob$discharge_disposition, n_enc, replace = TRUE, prob$p / sum(prob$p)))] # make sure probs add up to 1 (see addition of constant above)
+
+    ## Simulate LOS to derive ADMISSION_DATE_TIME
+    # create right-skewed distribution with randomly drawn offset by site
+    hosp_data[, los := rsn(n_enc, rnorm(1, 1.02, .05), .2, 10)^10]
+    hosp_data[, admission_date_time := format(as.POSIXct(ymd_hm(discharge_date_time) - ddays(los)), format = "%Y-%m-%d %H:%M")]
+
+    ## Alternate level of care (ALC) & days spent in ALC
+    # ALC flag
+    prob <- data.table(
+      "alc_service_transfer_flag" = c("FALSE", "TRUE", NA),
+      "p" = c(.85, .11, .04)
+    )
+    prob[, p := t(rdirichlet(1, alpha = prob$p / 0.05))] # 0.05 = level of variability
+    hosp_data[, alc_service_transfer_flag := sample(prob$alc_service_transfer_flag, n_enc, replace = TRUE, prob$p / sum(prob$p))] # make sure probs add up to 1 (see addition of constant above)
+
+    # Days spent in ALC (as integer)
+    # If ALC = FALSE, ALC days are either coded as 0 or NA (random across sites)
+    hosp_data[alc_service_transfer_flag == "FALSE", number_of_alc_days := sample(c(0, NA), 1, prob = c(.8, .2))]
+    # If ALC = TRUE, ALC days are drawn from uniform distribution between 0 and LOS (divided by 1.5 because ALC should be < LOS)
+    # Note: because ALC is rounded UP, this results in some entries where ALC > LOS (especially for cases with short LOS); this mimics entries we find in our real data as well
+    hosp_data[alc_service_transfer_flag == "TRUE", number_of_alc_days := ceiling(runif(.N, 0, ceiling(los / 1.5)))]
+    # for cases where number_of_alc_days != NA, alc_service_transfer_flag is NA anywhere from 0-100% by site (mostly 0 or 100, but some in-between), so let's mimic that
+    hosp_data[
+      genc_id %in% hosp_data[!is.na(number_of_alc_days)][
+        sample(.N, size = round(sample(c(0, .25, .50, .75, 1), prob = c(.59, .05, .05, .01, .3), 1) * .N)), "genc_id"
+      ],
+      alc_service_transfer_flag := NA
+    ]
+
+    # randomly recode values referring to FALSE/TRUE to simulate real messiness of ALC coding
+    coding <- t(
+      data.table(
+        code1 = c("FALSE", "TRUE"),
+        code2 = c("0", "1"),
+        code3 = c("0", "99"),
+        code4 = c("N", "Y"),
+        code5 = c("n", "y"),
+        code6 = c("false", "true"),
+        code7 = c("non-ALC", "ALC"),
+        code8 = c(NA, "Y")
+      ) # this is intentional, some sites only code "true", everything else is missing...
+    )
+    code <- sample(1:nrow(coding), 1)
+
+    hosp_data[alc_service_transfer_flag == FALSE, alc_service_transfer_flag := coding[code, 1]]
+    hosp_data[alc_service_transfer_flag == TRUE, alc_service_transfer_flag := coding[code, 2]]
+    # code missing as NA or "" (randomly per site)
+    hosp_data[is.na(alc_service_transfer_flag), alc_service_transfer_flag := sample(c(NA, ""), 1, prob = c(.8, .2))]
+
+    return(hosp_data)
+  }
+
+
+  # note: split data by hospital before running foverlaps to avoid working with massive tables
+  cohort_hospitals <- split(data, data$hospital_num)
+  data_all <- lapply(cohort_hospitals, add_vars)
+
+
+  ##  Combine all
+  data <- do.call(rbind, data_all)
+
+  ## Select relevant output variables
+  data <- data[order(genc_id), .(
+    genc_id,
+    hospital_num,
+    admission_date_time,
+    discharge_date_time,
+    age,
+    gender,
+    discharge_disposition,
+    alc_service_transfer_flag,
+    number_of_alc_days
+  )]
+
+  # Return as data.frame (instead of data.table) as this is what SQL queries return
+  data <- as.data.frame(data)
+
+  return(data)
+}
+
+
+
+#' @title
+#' Generated simulated lab data
+#'
+#' @description
+#' Designed to mimic the most important elements of the GEMINI lab table as defined in the
+#' [GEMINI Data Repository Dictionary](https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5).
+#'
+#' @param id (`numeric`)\cr
+#' A single identifier that is repeated to match the length of `value`.
+#'
+#' @param omop (`character`)\cr
+#' Codes corresponding to OMOP concept identifiers.
+#'
+#' @param value (`numeric`)\cr
+#' Simulated result values for each lab test measurement.
+#'
+#' @param unit (`character`)\cr
+#' Units corresponding to the particular lab test as defined by `omop`. It is repeated to match the length of `value`.
+#'
+#' @param mintime (`character`)\cr
+#' In the format yyyy-mm-dd hh:mm. Earliest recorded test performed time.
+#'
+#' @return (`data.table`)\cr
+#' With the columns, `id`, `omop`, `value`, `unit`, and `collection_date_time` as described above.
+#'
+#' @export
+#'
+#' @examples
+#' lab <- dummy_lab(1, 3024641, c(7, 8, 15, 30), "mmol/L", "2023-01-02 08:00")
+#'
+dummy_lab <- function(id, omop, value, unit, mintime) {
+  res <- data.table(
+    genc_id = rep(id, length(value)),
+    test_type_mapped_omop = omop,
+    result_value = value,
+    result_unit = rep(unit, length(value)),
+    collection_date_time = format(as.POSIXct(mintime, tz = "UTC") + sample(0:(24 * 60 * 60 - 1), size = length(value), replace = TRUE), "%Y-%m-%d %H:%M")
+  )
+  return(res)
+}
+
+
+#' @title
+#' Generated simulated administrative data
+#'
+#' @description
+#' Designed to partially mimic the `admdad` table as defined in the
+#' [GEMINI Data Repository Dictionary](https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5).
+#'
+#' @param id (`numeric`)\cr
+#' A single identifier that is repeated to match the length of `value`.
+#'
+#' @param admtime (`character`)\cr
+#' In the format yyyy-mm-dd hh:mm. Corresponds to the admission time of the encounter.
+#'
+#' @return (`data.table`)\cr
+#' With the columns `id` and `admission_date_time` as described above.
+#'
+#' @export
+#'
+#' @examples
+#' admdad <- dummy_admdad(1, "2023-01-02 00:00")
+#'
+dummy_admdad <- function(id, admtime) {
+  res <- data.table(
+    genc_id = id,
+    admission_date_time = format(as.POSIXct(admtime, tz = "UTC"), "%Y-%m-%d %H:%M")
+  )
+  return(res)
 }
 
 
@@ -486,21 +828,6 @@ NULL
 #'
 #' @keywords internal
 #' @name dummy_ccsr
-#' @docType data
-#'
-NULL
-
-#' @title
-#' Dummy ipadmdad data
-#'
-#' @description
-#' Used to generate a subset of variables from the ipadmdad table. Simulates
-#' clustering by hospital where patient populations have different
-#' characteristics at different sites (i.e., function can be used to simulate
-#' data for random intercept models).
-#'
-#' @keywords internal
-#' @name dummy_ipadmdad
 #' @docType data
 #'
 NULL